+++ packages/reactivity-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0
 
 -   New feature: `reactiveStruct`, a utility to create simple reactive classes.
+-   Fix: start the internal MessageChannel to deliver async messages
 
 ## v0.1.0